+++ src/main/java/br/gov/ufg/api/ApiApplication.java
@@ -1,0 +1,18 @@
+package br.gov.ufg.api;
+
+import br.gov.ufg.controller.ProdutoController;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootApplication
+@ComponentScan(basePackageClasses = ProdutoController.class)
+public class ApiApplication {
+
+	public static void main(String[] args) {
+
+		SpringApplication.run(ApiApplication.class, args);
+
+	}
+}
+
