+++ src/main/java/br/gov/ufg/entity/LoginRequest.java
@@ -1,0 +1,28 @@
+package br.gov.ufg.entity;
+
+public class LoginRequest {
+    String email;
+    String password;
+
+    public LoginRequest(String email, String password) {
+        this.email = email;
+        this.password = password;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+}
