+++ src/test/java/br/gov/ufg/api/ApiApplicationTests.java
@@ -1,0 +1,13 @@
+package br.gov.ufg.api;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
